Add method to identify compatible host
This commit adds a method that returns the least utilized compatible host for a given VM. Additionally a method is added to provide a capability to live migrate virtual machines to an alternate host. Without this change there is no way provided to identify a host suitable for live VM migration.