Check failed VMs for hostname configuration
This commit adds a capability to investigate a VM hostname configuration upon being evaluated for failure after timeout in order to determine whether the failure is resulting from a hostname configuration issue at deploy time. If the VM has reached timeout, the VM has an IP address, and DNS name does not equal the VM name, then the failure message is expanded to indicate that ddns configuration failed due to a failure of setting VM hostname configuration at deploy time. Without this change we only know that a VM fails, but gain no additional context.