Move host selection logic in vsphere.rb
This commit updates vsphere.rb to contain updates to host selection
logic there. Without this change changes to host selection that are
vsphere specific are polluting pool_manager which impacts the providers
concept being pluggable. Additionally, provider_hosts is moved to the
provider base class so it can be accessible on a per provider basis.