Add capability to update templates via config api
This commit adds the capability to update templates via the configuration API. Without this change templates can only be updated via a static configuration file. Once updated a template update is sticky and the redis value takes precedence over a value in a configuration file loaded at startup time.

When a template version is updated pool_manager will stop clone operations for that pool, remove all ready and pending instances, create template deltas for the new template, and then resume clone operations and fill the pool with clones built using the updated template version.

Additionally, the method used to detect a VM template is updated to simplify the number of objects looked up and the method used to lookup those objects.